% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_data.R
\name{getSensorData}
\alias{getSensorData}
\title{Load and parse real field data from github storage
Разбираемся с проблемой отображения UTF}
\usage{
getSensorData()
}
\description{
Load and parse real field data from github storage
Разбираемся с проблемой отображения UTF
}

