% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeframe.R
\name{getTimeframe}
\alias{getTimeframe}
\title{Get appropriate timeframe window from current time.
Timeframe adjusted up to day unit and has forward and backward extension.}
\usage{
getTimeframe(days_back = 7, days_forward = 3)
}
\description{
Get appropriate timeframe window from current time.
Timeframe adjusted up to day unit and has forward and backward extension.
}

