% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_groups.R
\name{tgroups}
\alias{hgroup.enum}
\alias{hgroup.enum0}
\alias{hgroup.enum1}
\alias{tgroups}
\title{Different functions to group measurements with seconds resolution into high ordered time intervals}
\usage{
hgroup.enum0(date, time.bin = 4)

hgroup.enum1(date, time.bin = 4)

hgroup.enum(date, time.bin = 4)
}
\description{
Different functions to group measurements with seconds resolution into high ordered time intervals
}

