% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate_raw_weather_data.R
\name{opweather}
\alias{calcRainPerDate}
\alias{extractWeather}
\alias{gatherRawWeatherData}
\alias{opweather}
\title{Load, calc, combine and filter history weather data & prediction}
\usage{
gatherRawWeatherData()

extractWeather(raw_weather, timeframe)

calcRainPerDate(raw_weather)
}
\description{
Load, calc, combine and filter history weather data & prediction
}

